ProblemJsonのリファクタリング
